requirements.txt
azure-cognitiveservices-vision-customvision
Pillow
PyYAML